Move Common Methods Out of GuiPluginTools
